requirements.txt
selenium
pytest
pytest-xdist
pytest-rerunfailures
webdriver-manager
python-dotenv
pyyaml
pykwalify
allure-pytest